add application ps api for openshift
