+++ ai-services/internal/pkg/catalog/apiserver/repository/application_service/openshift.go
@@ -7,6 +7,7 @@ import (
 	"github.com/google/uuid"
 	apimodels "github.com/project-ai-services/ai-services/internal/pkg/catalog/apiserver/models"
 	"github.com/project-ai-services/ai-services/internal/pkg/catalog/types"
+	catalogutils "github.com/project-ai-services/ai-services/internal/pkg/catalog/utils"
 	runtimeTypes "github.com/project-ai-services/ai-services/internal/pkg/runtime/types"
 )
 
@@ -32,8 +33,9 @@ func (s *OpenShiftApplicationService) GetApplicationResources(_ context.Context,
 	return nil, errOpenShiftNotSupported
 }
 
-func (s *OpenShiftApplicationService) ApplicationsPs(_ context.Context, _ uuid.UUID) (*types.ApplicationPSResponse, error) {
-	return nil, errOpenShiftNotSupported
+// ApplicationsPs retrieves pod/container status by querying the application's OpenShift namespace.
+func (s *OpenShiftApplicationService) ApplicationsPs(ctx context.Context, appID uuid.UUID) (*types.ApplicationPSResponse, error) {
+	return s.ApplicationServiceBase.ApplicationsPs(ctx, appID, catalogutils.AppNamespace(appID))
 }
 
 // Made with Bob
